Tidy variables. Move isBufferBinary to utils.
